Add SparkMergeTreeSink
